[Feat] 태그 처리 개선 및 토큰 검증 로직 정제
존재하지 않는 태그를 데이터베이스에 저장하고 이를 게시물과 연동하는 태그 생성 프로세스를 개선했습니다. 또한, JwtProvider에서 null 토큰을 보다 효과적으로 처리할 수 있도록 토큰 검증 로직을 정제했습니다.

+++ src/main/java/com/scout/tech/config/JwtProvider.java
@@ -48,9 +48,11 @@ public class JwtProvider {
 
     public boolean verifyRefreshTokenWithRedis(Long userId, String token) {
         RefreshToken refreshToken = refreshTokenRepository.findById(userId).orElseThrow(TokenInvalidException::new);
-        if (!refreshToken.getToken().equals(token)) {
-            refreshTokenRepository.deleteById(userId);
-            return false;
+        if (!(token == null)) {
+            if (!refreshToken.getToken().equals(token)) {
+                refreshTokenRepository.deleteById(userId);
+                return false;
+            }
         }
         return true;
     }
